expr.{cpp,hpp}: change semantics of expr pieces so that they do include the piece op
Not knowing the piece op was making it impossible to determine the
byte/bit length of a piece, or to know whether to apply the "all bytes"
fallback that comes with implicit pieces. The semantics of copy() do
change unfortunately, so we now have also copy_as_if_whole() which
strips the piece op.